Moved game types to their own file
